DRY implementation of Protobuf JVM symbol mapping
